Record a reload of an expired entry as a climber miss
A computeIfAbsent or compute that finds the entry expired, or its value
collected, loads into the node that is already linked, so the policy
receives an UpdateTask where a reaped entry would have produced an
AddTask. Credited as an ordinary access, that recorded a climber hit for
an operation the application experienced as a miss, and left the sampled
rate turning on whether maintenance reaped the entry before the reload
arrived. CacheStats was correct throughout; only the sample was wrong.

The level error is not what makes it matter. A steady contamination
cancels in every difference the machine takes, the reactive law's rate
change and the walk, veto, and anchor comparisons alike, since each
subtracts one rate from another. What has nothing to subtract it against
is the density tier's within-sample ratio, which divides each region's
hits by that region's capacity. On a seeded synthetic reuse stream at a
maximum of 512, with the reuse gap set to the expiration duration, the
mislabel moved the converged window from 402 entries to 5 on every seed
and at both maintenance lags. Nearly every phantom lands in main, since
an entry that survives to expire is one the main space is holding, so
the density law reads main as earning and steers capacity out of the
window. The reload rate is itself a function of the window, so the error
feeds itself: the collapsed window took 13.6% of requests as reloads
against the corrected window's 4.2%.

Access replaces the quietly flag. A reload keeps the sketch increment an
insertion performs, and the task rather than the access records the
miss, so an entry too heavy to reorder still pays it. Suppressing the
access with quietly instead would drop the increment AddTask performs. A
cache with neither expiration nor reference values cannot reach the
branch, so the simulator's policies are bit-identical under this and the
gate battery cannot price it. A put over an expired entry takes the same
in-place path and is still credited a hit, being a write not a load.

+++ caffeine/src/testFixtures/java/com/github/benmanes/caffeine/cache/ClimberInvariants.java
@@ -15,14 +15,14 @@
  */
 package com.github.benmanes.caffeine.cache;
 
+import static com.github.benmanes.caffeine.cache.WindowClimber.RETREAT_COVER;
+import static com.github.benmanes.caffeine.cache.WindowClimber.Anchor.RETEST_SETTLE;
 import static com.github.benmanes.caffeine.cache.WindowClimber.AuditClock.AUDIT_WAIT_FIRST;
 import static com.github.benmanes.caffeine.cache.WindowClimber.AuditClock.AUDIT_WAIT_INITIAL;
 import static com.github.benmanes.caffeine.cache.WindowClimber.AuditClock.AUDIT_WAIT_MAX;
-import static com.github.benmanes.caffeine.cache.WindowClimber.Anchor.RETEST_SETTLE;
 import static com.github.benmanes.caffeine.cache.WindowClimber.Ladder.PROBE_BACKOFF_INITIAL;
 import static com.github.benmanes.caffeine.cache.WindowClimber.Ladder.PROBE_BACKOFF_MAX;
 import static com.github.benmanes.caffeine.cache.WindowClimber.Ladder.PROBE_CRASH_ESCALATION;
-import static com.github.benmanes.caffeine.cache.WindowClimber.RETREAT_COVER;
 import static com.github.benmanes.caffeine.cache.WindowClimber.Reading.MAX_STEP_FRACTION;
 import static com.github.benmanes.caffeine.cache.WindowClimber.Step.MIN_INITIAL_STEP;
 import static com.github.benmanes.caffeine.cache.WindowClimber.Walk.AUDIT_CRASH_PERSISTENCE;
@@ -169,7 +169,7 @@ final class ClimberInvariants {
     // of headroom. The bound is stricter than the adjudicated reality: a negative policyWeight
     // inflates the transfer loop's quota and carries the excess back, a sanctioned transient that
     // no subject-validated test drives.
-    long bound = Math.max((long) MIN_INITIAL_STEP, (long) (MAX_STEP_FRACTION * maximum)) + 1;
+    long bound = (long) Math.max(MIN_INITIAL_STEP, (MAX_STEP_FRACTION * maximum)) + 1;
     assertWithMessage("the adjustment honors the maximum step")
         .that(Math.abs(climber.adjustment())).isAtMost(bound);
     assertWithMessage("the step size honors the maximum step")
